+++ project-base/src/Shopsys/ShopBundle/Model/Mail/Exception/EmptyMailException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Mail\Exception;
+
+use Exception;
+
+class EmptyMailException extends Exception implements MailException
+{
+}