+++ project-base/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - cache is cleared before PHPUnit tests only when run via [Phing targets](docs/introduction/phing-targets.md), not when run using `phpunit` directly (@PetrHeinz)
 - PHPUnit tests now fail on warning (@TomasLudvik)
 - end of support of PHP 7.0 (@TomasLudvik)
+- emails with empty subject or body are no longer sent (@stanoMilan)
 
 ### Fixed
 - `BrandFacade::create()` now generates friendly URL for all domains (@sspooky13)